separated move project resources actions and modal
